refactor: 중복코드 제거를 위한 DaoConfiguration 도입
