+++ src/main/java/dao/GameDao.java
@@ -1,29 +1,15 @@
 package dao;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import model.Team;
 
 public class GameDao {
 
-    private static final String SERVER = "localhost:13306";
-    private static final String DATABASE = "janggi";
-    private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
-    private static final String USERNAME = "root";
-    private static final String PASSWORD = "root";
+    private final DaoConfiguration daoConfiguration;
 
-    public Connection getConnection() {
-        // 드라이버 연결
-        try {
-            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION,
-                USERNAME, PASSWORD);
-        } catch (final SQLException e) {
-            System.err.println("DB 연결 오류:" + e.getMessage());
-            e.printStackTrace();
-            return null;
-        }
+    public GameDao(DaoConfiguration daoConfiguration) {
+        this.daoConfiguration = daoConfiguration;
     }
 
     /***
@@ -33,7 +19,7 @@ public class GameDao {
      */
     public void addTurn(Team turn) {
         final var query = "INSERT INTO GAME(turn) VALUES (?)";
-        try (final var connection = getConnection();
+        try (final var connection = daoConfiguration.getConnection();
             final var preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setString(1, turn.name());
             preparedStatement.execute();
@@ -44,7 +30,7 @@ public class GameDao {
 
     public boolean deleteTurn() {
         final var query = "DELETE FROM GAME";
-        try (final var connection = getConnection();
+        try (final var connection = daoConfiguration.getConnection();
             final var preparedStatement = connection.prepareStatement(query)) {
             return preparedStatement.execute();
         } catch (final SQLException e) {
@@ -54,7 +40,7 @@ public class GameDao {
 
     public boolean updateTurn(Team turn) {
         final var query = "UPDATE GAME SET turn = ? WHERE game_id = 1";
-        try (final var connection = getConnection();
+        try (final var connection = daoConfiguration.getConnection();
             final var preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setString(1, turn.name());
             return preparedStatement.execute();
@@ -63,16 +49,15 @@ public class GameDao {
         }
     }
 
-
-
     /***
      * 현재 상황에서는 단일 게임만 존재함
      * 그렇기에 game_id를 1로 고정 함.
      * 추후 다른 게임이 추가된다면 해당 코드의 game_id = 1 부분을 수정해야 함.
      */
+
     public Team getTurn() {
         final var query = "SELECT turn FROM GAME WHERE game_id = ?";
-        try (final var connection = getConnection();
+        try (final var connection = daoConfiguration.getConnection();
             final var preparedStatement = connection.prepareStatement(query)) {
             preparedStatement.setInt(1, 1); // 현 시점에서는 game_id를 1로 고정한다.
             ResultSet resultSet = preparedStatement.executeQuery();
